+++ scripts/package.sh
@@ -2,14 +2,29 @@
 
 set -e
 
-version=`git describe --dirty --tags || echo unknown`
+dev=""
+if [ "$1" = '-d' ]; then
+    dev=":dev"
+fi
 
-npm run build
-mkdir -p packages
-cp -r vector vector-$version
+if [ -n "$DIST_VERSION" ]; then
+    version=$DIST_VERSION
+else
+    version=`git describe --dirty --tags || echo unknown`
+fi
+
+npm run clean
+npm run build$dev
+
+# include the sample config in the tarball. Arguably this should be done by
+# `npm run build`, but it's just too painful.
+cp config.sample.json webapp/
+
+mkdir -p dist
+cp -r webapp vector-$version
 echo $version > vector-$version/version
-tar chvzf packages/vector-$version.tar.gz vector-$version
+tar chvzf dist/vector-$version.tar.gz vector-$version
 rm -r vector-$version
 
 echo
-echo "Packaged packages/vector-$version.tar.gz"
+echo "Packaged dist/vector-$version.tar.gz"