+++ webpack.config.js
@@ -1,8 +1,21 @@
 var path = require('path');
 var webpack = require('webpack');
 var ExtractTextPlugin = require("extract-text-webpack-plugin");
+var HtmlWebpackPlugin = require('html-webpack-plugin');
 
 module.exports = {
+    entry: {
+        "bundle": "./src/vector/index.js",
+
+        // We ship olm.js as a separate lump of javascript. This makes it get
+        // loaded via a separate <script/> tag in index.html (which loads it
+        // into the browser global `Olm`), and define it as an external below.
+        //
+        // (we should probably make js-sdk load it asynchronously at some
+        // point, so that it doesn't block the pageload, but that is a separate
+        // problem)
+        "olm": "olm/olm.js",
+    },
     module: {
         preLoaders: [
             { test: /\.js$/, loader: "source-map-loader" }
@@ -19,9 +32,15 @@ module.exports = {
             // there is no need for webpack to parse them - they can just be
             // included as-is.
             /highlight\.js\/lib\/languages/,
+
+            // olm takes ages for webpack to process, and it's already heavily
+            // optimised, so there is little to gain by us uglifying it.
+            /olm\/(javascript\/)?olm\.js$/,
         ],
     },
     output: {
+        path: path.join(__dirname, "webapp"),
+        filename: "[name].[chunkhash].js",
         devtoolModuleFilenameTemplate: function(info) {
             // Reading input source maps gives only relative paths here for
             // everything. Until I figure out how to fix this, this is a
@@ -38,7 +57,8 @@ module.exports = {
         alias: {
             // alias any requires to the react module to the one in our path, otherwise
             // we tend to get the react source included twice when using npm link.
-            react: path.resolve('./node_modules/react'),
+            "react": path.resolve('./node_modules/react'),
+            "react-dom": path.resolve('./node_modules/react-dom'),
             "react-addons-perf": path.resolve('./node_modules/react-addons-perf'),
 
             // same goes for js-sdk
@@ -46,12 +66,10 @@ module.exports = {
         },
     },
     externals: {
-        // olm takes ages for webpack to process, and it's already heavily
-        // optimised, so there is little to gain by us uglifying it. We
-        // therefore use it via a separate <script/> tag in index.html (which
-        // loads it into the browser global `Olm`), and reference it as an
-        // external here.
         "olm": "Olm",
+        // Don't try to bundle electron: leave it as a commonjs dependency
+        // (the 'commonjs' here means it will output a 'require')
+        "electron": "commonjs electron",
     },
     plugins: [
         new webpack.DefinePlugin({
@@ -60,9 +78,37 @@ module.exports = {
             }
         }),
 
-        new ExtractTextPlugin("bundle.css", {
-            allChunks: true
+        new ExtractTextPlugin(
+            "[name].[contenthash].css",
+            {
+                allChunks: true
+            }
+        ),
+
+        new HtmlWebpackPlugin({
+            template: './src/vector/index.html',
+
+            // we inject the links ourselves via the template, because
+            // HtmlWebpackPlugin wants to put the script tags either at the
+            // bottom of <head> or the bottom of <body>, and I'm a bit scared
+            // about moving them.
+            inject: false,
         }),
     ],
-    devtool: 'source-map'
+    devtool: 'source-map',
+
+    // configuration for the webpack-dev-server
+    devServer: {
+        // serve unwebpacked assets from webapp.
+        contentBase: './webapp',
+    },
 };
+
+// olm is an optional dependency. Ignore it if it's not installed, to avoid a
+// scary-looking error.
+try {
+    require('olm');
+} catch (e) {
+    console.log("Olm is not installed; not shipping it");
+    delete(module.exports.entry["olm"]);
+}